% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bedgraph.to.depth}
\alias{bedgraph.to.depth}
\title{Convert a BedGraph fiel to a vector of depth-like signals}
\usage{
bedgraph.to.depth(inFpath, chr, start, end, reverse = FALSE)
}
\arguments{
\item{inFpath}{character, the path of a bedgraph file}

\item{chr}{character, the chromosome}

\item{start}{numeric, the start position (0-base)}

\item{end}{numeric, the end position}

\item{reverse}{logical, whether a reverse strand}
}
\value{
numeric
}
\description{
Convert a BedGraph fiel to a vector of depth-like signals
}
\author{
Xiaobei Zhao
}

