% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tbl.R
\name{doc_tbl}
\alias{doc_tbl}
\title{Get the a table of Google Drive documents}
\usage{
doc_tbl(token = token)
}
\arguments{
\item{token}{your google authentication token}
}
\value{
tibble containing id, title, and date created of documents in google drive
}
\description{
Get the a table of Google Drive documents
}
\examples{
\dontrun{
gd_tbl <- doc_tbl(token = token)
}
}

