% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Get the ids of Google Drive documents}
\usage{
get_id(edit_url)
}
\arguments{
\item{edit_url}{the url you see when editing your Google Document}
}
\value{
your document id, to be input in `GMD()`
}
\description{
Get the ids of Google Drive documents
}
\examples{
edit_url <- "https://docs.google.com/document/d/1RTCQ67mpZTKe9ddllVNCBom5uC2KMFjktKHb1mjWKOM/edit"
id <- get_id(edit_url)

}

