% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_update.R
\name{live_update}
\alias{live_update}
\title{Setup a constant live pulling from the google doc specified.}
\usage{
live_update(document, refresh_rate = 1)
}
\arguments{
\item{document}{The return from running `GMD()`. E.g. `my_doc <- GMD(token, url)`}

\item{refresh_rate}{How many seconds you want to wait before re-pulling google doc.}
}
\value{
Downloads from google doc and makes into an RMD continuously a `.Rmd`.
}
\description{
Setup a constant live pulling from the google doc specified.
}
\examples{
\dontrun{
my_doc <- GMD(token, url)
live_update(my_doc)
}
}

