\name{print.gmd}
\alias{print.gmd}
\title{Print Function for Class gmd}
\usage{print.gmd(x, print.mode=c("brief", "full"), digits=3, ...)}
\description{Print Function for Class \code{gmd}}
\details{Print Function for Class \code{gmd}}
\seealso{\code{\link{gmd}}}
\references{See \code{help(GMD)}}
\arguments{\item{x}{an object of class \code{gmd}.}
\item{print.mode}{a string, indicating whether to print in \emph{full} mode (\emph{default}).}
\item{digits}{integer, indicating the number of decimal places to be printed.}
\item{...}{arguments to be passed to methods, see \code{print}.}}
