% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_doc.R
\name{render_doc}
\alias{render_doc}
\title{Render Google Document as html or pdf}
\usage{
render_doc(doc, output_format = "html_document")
}
\arguments{
\item{doc}{function to pull Google document obtained from `GMD()`}

\item{output_format}{"html_document" or "pdf_document"}
}
\value{

}
\description{
Render Google Document as html or pdf
}
\examples{

}

