\name{summary.gmd}
\alias{summary.gmd}
\title{Summary Function for Class gmd}
\usage{summary.gmd(object, ...)}
\description{Summary Function for Class \code{gmd}}
\details{Summary Function for Class \code{gmd}}
\seealso{\code{\link{gmd}}}
\references{See \code{help(GMD)}}
\arguments{\item{object}{an object of class \code{gmd}.}
\item{...}{arguments to be passed to methods, see \code{summary}.}}
