% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMD.R
\name{GMD}
\alias{GMD}
\title{Create an R Markdown document from a Google Document}
\usage{
GMD(doc, token)
}
\arguments{
\item{doc}{your google documents unique ID or edit URL. The unique ID can be obtained by using the `doc_tbl()` function and navigating to your document of interest. The edit URL is the Google Doc url you use to edit the document.}

\item{token}{your google authentication token}
}
\value{
creates a `.Rmd` file in your working directory and, if specified, renders the document
}
\description{
Create an R Markdown document from a Google Document
}
\examples{
\dontrun{
edit_url <- "https://docs.google.com/document/d/1RTCQ67mpZTKe9ddllVNCBom5uC2KMFjktKHb1mjWKOM/edit"
GMD(doc = edit_url, token)
}
}

